Fixed fmod() versus fmodf() usage in flight/...

+++ flight/Modules/Sensors/sensors.c
@@ -68,7 +68,7 @@
 #define SENSOR_PERIOD 2
 
 #define F_PI 3.14159265358979323846f
-#define PI_MOD(x) (fmod(x + F_PI, F_PI * 2) - F_PI)
+#define PI_MOD(x) (fmodf(x + F_PI, F_PI * 2) - F_PI)
 // Private types
 
 // Private variables
